📝 docs(what-is-programming): add content explaining the concepts of computer and program
This commit adds content to the "What is Programming?" page, explaining the concepts of a computer and a program. It provides an overview of the components of a computer, including the processor (CPU) and RAM, and their roles in executing instructions and storing data. It also defines a program as the software that instructs a computer on what tasks to perform and describes the different components of a program,